Fix activation function
Before the fix, the activation function generated a few "no activation"
states during the 1ms interval between EMG samples, although the EMG
signal was above the activation threshold. To fix this, an activation
variable "isActive" has been added. This variable is returned at the end
of the activation function and its value is modified only if the signal
is above or below the activation threshold.

Also, the "0" and "1" values of the "sampleOk" variable have been
changed for "false" and "true" values.